Rename Events and Filters.
