+++ PokeAlarm/Events/BaseEvent.py
