+++ PokeAlarm/Filters/BaseFilter.py
