+++ PokeAlarm/Filters/__init__.py
@@ -1,6 +1,6 @@
-from Filter import Filter
-from Monster import Monster
-from Stop import Stop
-from Gym import Gym
-from Egg import Egg
-from Raid import Raid
+from BaseFilter import Filter
+from MonFilter import Monster
+from StopFilter import Stop
+from GymFilter import Gym
+from EggFilter import Egg
+from RaidFilter import Raid
